split out browser plugin
